requirements.txt
discord.py
python-dotenv
supabase
pangea-sdk